Create generic forms file, refactor documentation
