+++ cli/node_account/help.py
@@ -2,6 +2,6 @@
 Provide help messages for command line interface's node account commands.
 """
 NODE_ACCOUNT_ADDRESS_ARGUMENT_HELP_MESSAGE = 'Node account address to get information about node account by.'
-NODE_ACCOUNT_ADDRESS_TO_ARGUMENT_HELP_MESSAGE = 'Node account address to transfer tokens to.'
+ACCOUNT_ADDRESS_TO_ARGUMENT_HELP_MESSAGE = 'Account address to transfer tokens to.'
 AMOUNT_ARGUMENT_HELP_MESSAGE = 'Amount to transfer.'
-PRIVATE_KEY_ARGUMENT_HELP_MESSAGE = 'Node account\'s private key to transfer tokens from.'
+PRIVATE_KEY_ARGUMENT_HELP_MESSAGE = 'Account\'s private key to transfer tokens from.'
